file restructure and renaming
