+++ src-ui/renderer/reactive/footer/footer.spec.ts
