+++ src-ui/renderer/reactive/footer/index.ts
@@ -1,0 +1,1 @@
+export { footer } from './footer'
