+++ src-ui/renderer/reactive/index.ts
@@ -1,11 +1,4 @@
-/**
- * renderer/reactive
- * State-based components that
- * re-render independently based on events
- */
 export { footer } from './footer'
-export { renderRemoteDbLogs } from './remote-db-logs'
-export { renderRemoteDbSetupDialog } from './remote-db-setup-dialog'
-export { renderSidebarCreateNote } from './sidebar-create-note'
-export { renderSidebarNoteList } from './sidebar-note-list'
+export { renderSidebarCreateNote, renderSidebarNoteList } from './sidebar'
+export { renderRemoteDbLogs, renderRemoteDbDialog } from './remote-db-dialog'
 export { renderNoteDetailsDialog } from './note-details-dialog'
