+++ src-ui/renderer/reactive/note-details-dialog/index.ts
@@ -1,0 +1,1 @@
+export { renderNoteDetailsDialog } from './note-details-dialog'
