+++ src-ui/renderer/reactive/note-details-dialog/note-details-dialog.css
