+++ src-ui/renderer/reactive/remote-db-dialog/index.ts
@@ -1,0 +1,2 @@
+export { renderRemoteDbLogs } from './remote-db-logs'
+export { renderRemoteDbDialog } from './remote-db-dialog'
