+++ src-ui/renderer/reactive/sidebar/index.ts
@@ -1,0 +1,2 @@
+export { renderSidebarCreateNote } from './sidebar-create-note'
+export { renderSidebarNoteList } from './sidebar-note-list'
