+++ src-ui/renderer/reactive/sidebar/sidebar-create-note.spec.ts
