+++ src-ui/renderer/reactive/sidebar/sidebar-note-list.spec.ts
